Move scan executors
