+++ src/coprocessor/dag/batch_executor/scan/scan_executor.rs
@@ -17,7 +17,7 @@ use tipb::schema::ColumnInfo;
 
 use crate::storage::{Key, Store};
 
-use super::interface::*;
+use super::super::interface::*;
 use super::ranges_iter::{PointRangePolicy, RangesIterator};
 use crate::coprocessor::codec::batch::LazyBatchColumnVec;
 use crate::coprocessor::dag::expr::EvalContext;
