+++ src/coprocessor/dag/batch_executor/scan/table_scan_executor.rs
@@ -21,7 +21,7 @@ use tipb::schema::ColumnInfo;
 use crate::storage::Store;
 use crate::util::collections::HashMap;
 
-use super::interface::*;
+use super::super::interface::*;
 use crate::coprocessor::codec::batch::{LazyBatchColumn, LazyBatchColumnVec};
 use crate::coprocessor::dag::expr::{EvalConfig, EvalContext};
 use crate::coprocessor::dag::Scanner;
